Ugly commit because I screwed up file renaming
Do we really need to talk about this? Okay, well in trying to make the files here easier to understand I renamed some files but I neglected to do it in git and then I pushed, so the only way to fix this (easily for me) is to delete the old files. I added a note to each renamed-file so you can match them up if you need to. Ugly! I know.